LayoutTests/http/tests/security/appcache-switching-private-browsing-expected.txt

CONSOLE MESSAGE: line 8: Test started
CONSOLE MESSAGE: line 1: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: line 18: PASS
This test passes if the application cache does not load while in private browsing



--------
Frame: '<!--frame1-->'
--------
